refactor: decorate connection of modules
